local currency label positioned by autolayout

+++ BreadWallet/BRSettingsViewController.m
@@ -465,7 +465,7 @@
         case 4:
             return NSLocalizedString(@"fees are only optional for high priority transactions", nil);
 
-        default: NSAssert(FALSE, @"%s:%d %s: unkown section %d", __FILE__, __LINE__,  __func__, section);
+        default: NSAssert(FALSE, @"%s:%d %s: unkown section %d", __FILE__, __LINE__,  __func__, (int)section);
     }
     
     return nil;
